+++ examples/04_typeahead/src/App.tsx
@@ -2,18 +2,25 @@ import * as React from 'react';
 
 import GitHubSearch from './GitHubSearch';
 
-const { useState } = React;
+const {
+  useState,
+  StrictMode,
+  // @ts-ignore
+  unstable_ConcurrentMode: ConcurrentMode,
+} = React;
 
 const App = () => {
   const [query, setQuery] = useState('');
   return (
-    <React.StrictMode>
-      <div>
-        Query:
-        <input value={query} onChange={e => setQuery(e.target.value)} />
-        {query && <GitHubSearch query={query} />}
-      </div>
-    </React.StrictMode>
+    <StrictMode>
+      <ConcurrentMode>
+        <div>
+          Query:
+          <input value={query} onChange={e => setQuery(e.target.value)} />
+          {query && <GitHubSearch query={query} />}
+        </div>
+      </ConcurrentMode>
+    </StrictMode>
   );
 };