+++ examples/07_race/src/App.tsx
@@ -2,12 +2,20 @@ import * as React from 'react';
 
 import DelayedData from './DelayedData';
 
+const {
+  StrictMode,
+  // @ts-ignore
+  unstable_ConcurrentMode: ConcurrentMode,
+} = React;
+
 const App = () => (
-  <React.StrictMode>
-    <div>
-      <DelayedData />
-    </div>
-  </React.StrictMode>
+  <StrictMode>
+    <ConcurrentMode>
+      <div>
+        <DelayedData />
+      </div>
+    </ConcurrentMode>
+  </StrictMode>
 );
 
 export default App;