+++ examples/08_wasm/src/App.tsx
@@ -2,19 +2,26 @@ import * as React from 'react';
 
 import CalcFib from './CalcFib';
 
-const { useState } = React;
+const {
+  useState,
+  StrictMode,
+  // @ts-ignore
+  unstable_ConcurrentMode: ConcurrentMode,
+} = React;
 
 const App = () => {
   const [count, setCount] = useState(1);
   return (
-    <React.StrictMode>
-      <div>
-        count:{count}
-        <button type="button" onClick={() => setCount(count + 1)}>+1</button>
-        <button type="button" onClick={() => setCount(count - 1)}>-1</button>
-        <CalcFib count={count} />
-      </div>
-    </React.StrictMode>
+    <StrictMode>
+      <ConcurrentMode>
+        <div>
+          count:{count}
+          <button type="button" onClick={() => setCount(count + 1)}>+1</button>
+          <button type="button" onClick={() => setCount(count - 1)}>-1</button>
+          <CalcFib count={count} />
+        </div>
+      </ConcurrentMode>
+    </StrictMode>
   );
 };