+++ src/use-async-task.js
@@ -1,5 +1,5 @@
 import {
-  useEffect,
+  useLayoutEffect,
   useReducer,
   useRef,
 } from 'react';
@@ -50,24 +50,34 @@ const createTask = (func, notifyUpdate) => {
 
 export const useAsyncTask = (func, inputs) => {
   const forceUpdate = useForceUpdate();
-  const task = useRef(null);
+  // inputs
   const prevInputs = useRef();
-  if (!prevInputs.current || !shallowArrayEqual(prevInputs.current, inputs)) {
+  useLayoutEffect(() => {
     prevInputs.current = inputs;
-    task.current = createTask(func, (updatedTask) => {
+  });
+  // task
+  const task = useRef(null);
+  let currentTask = task.current;
+  useLayoutEffect(() => {
+    // We need to set task.current before event hander can be called.
+    task.current = currentTask;
+    const cleanup = () => {
+      task.current = null;
+    };
+    return cleanup;
+  });
+  // create task
+  if (!currentTask || !shallowArrayEqual(prevInputs.current, inputs)) {
+    currentTask = createTask(func, (updatedTask) => {
+      // Note: task.start() should be called in useEffect or event handler,
+      // otherwise the task will be not updated.
       if (task.current && task.current.taskId === updatedTask.taskId) {
         task.current = updatedTask;
         forceUpdate();
       }
     });
   }
-  useEffect(() => {
-    const cleanup = () => {
-      task.current = null;
-    };
-    return cleanup;
-  }, []);
-  return task.current;
+  return currentTask;
 };
 
 export default useAsyncTask;